@@ -23,7 +23,7 @@
     "chat": "https://discord.link/twill"
   },
   "require": {
-    "php": "^7.1|^8.0",
+    "php": "^8.0",
     "ext-json": "*",
     "ext-pdo": "*",
     "astrotomic/laravel-translatable": "^11.5",
@@ -49,7 +49,7 @@
     "kalnoy/nestedset": "^5.0|^6.0",
     "nunomaduro/collision": "^3.0|^4.2|^5.0|^6.0",
     "orchestra/testbench": "~3.3|~3.4|~3.5|~3.6|~3.7|~3.8|^4.0|^5.0|^6.0|^7.0",
-    "nunomaduro/larastan": "^1.0",
+    "nunomaduro/larastan": "^1.0|^2.0",
     "phpunit/phpunit": "~5.0|~6.0|~7.0|~8.0|~9.0",
     "squizlabs/php_codesniffer": "*"
   },
