@@ -2,13 +2,13 @@
 
 namespace A17\Twill\Commands;
 
+use A17\Twill\Models\User;
+use A17\Twill\Services\Capsules\Manager;
 use Exception;
 use GuzzleHttp\Client;
-use A17\Twill\Models\User;
 use Illuminate\Support\Facades\Config;
-use Illuminate\Support\Str;
 use Illuminate\Support\Facades\Hash;
-use A17\Twill\Services\Capsules\Manager;
+use Illuminate\Support\Str;
 
 class CapsuleInstall extends Command
 {
@@ -30,7 +30,6 @@ class CapsuleInstall extends Command
      *
      * @var string
      */
-
     protected $description = 'Install a Twill Capsule';
 
     protected $repositoryUri;
@@ -40,6 +39,10 @@ class CapsuleInstall extends Command
     protected $capsuleName;
 
     protected $repositoryUrl;
+
+    protected string $namespace;
+
+    protected Manager $manager;
 
     /**
      * Create a new console command instance.
@@ -72,7 +75,7 @@ class CapsuleInstall extends Command
      */
     public function handle()
     {
-        if (!$this->checkParameters()) {
+        if (! $this->checkParameters()) {
             return 255;
         }
 
@@ -87,7 +90,7 @@ class CapsuleInstall extends Command
 
     protected function checkParameters()
     {
-        if (!$this->option('require') && !$this->option('copy')) {
+        if (! $this->option('require') && ! $this->option('copy')) {
             $this->error('Missing mandatory strategy: --require or --copy.');
 
             return false;
@@ -113,7 +116,7 @@ class CapsuleInstall extends Command
         } else {
             $capsule = Str::snake(Str::kebab($capsule));
 
-            if (!Str::contains($capsule, '/')) {
+            if (! Str::contains($capsule, '/')) {
                 $capsule = $this->getRepositoryPrefix() . "-$capsule";
             }
 
@@ -255,7 +258,7 @@ class CapsuleInstall extends Command
 
         $this->comment('');
 
-        if (!$installed) {
+        if (! $installed) {
             $this->error('Your capsule was not installed.');
         } else {
             $this->comment('Your capsule was installed successfully!');
@@ -276,7 +279,7 @@ class CapsuleInstall extends Command
 
     protected function download()
     {
-        if (!$this->cleanTempFile() || !$this->repositoryExists()) {
+        if (! $this->cleanTempFile() || ! $this->repositoryExists()) {
             return false;
         }
 
@@ -359,7 +362,7 @@ class CapsuleInstall extends Command
     {
         $return = shell_exec('which unzip');
 
-        return !empty($return);
+        return ! empty($return);
     }
 
     protected function unzipWithExtension($zip, $directory)
@@ -378,7 +381,7 @@ class CapsuleInstall extends Command
 
         unlink($zip);
 
-        if (!$success) {
+        if (! $success) {
             $this->error("Cound not read zip file: $zip");
 
             return false;
@@ -409,7 +412,7 @@ class CapsuleInstall extends Command
 
     public function makeDir($path)
     {
-        if (!file_exists($path)) {
+        if (! file_exists($path)) {
             mkdir($path, 0775, true);
         }
     }
