@@ -1,9 +1,10 @@
 <?php
 
 use Illuminate\Support\Facades\Request;
+use Illuminate\Support\Facades\Route;
 use Illuminate\Support\Str;
 
-if (!function_exists('moduleRoute')) {
+if (! function_exists('moduleRoute')) {
     /**
      * @param string $moduleName
      * @param string $prefix
@@ -29,7 +30,7 @@ if (!function_exists('moduleRoute')) {
         if (
             config('twill.allow_duplicates_on_route_names', true) ||
             ($prefix !== $moduleName &&
-                !Str::endsWith($prefix, '.' . $moduleName))
+                ! Str::endsWith($prefix, '.' . $moduleName))
         ) {
             $routeName .= "{$moduleName}";
         }
@@ -42,7 +43,7 @@ if (!function_exists('moduleRoute')) {
     }
 }
 
-if (!function_exists('getNavigationUrl')) {
+if (! function_exists('getNavigationUrl')) {
     /**
      * @param array $element
      * @param string $key
@@ -56,27 +57,22 @@ if (!function_exists('getNavigationUrl')) {
 
         if ($isModule) {
             $action = $element['route'] ?? 'index';
+
             return moduleRoute($key, $prefix, $action);
         } elseif ($isSingleton) {
             return moduleRoute($key, $prefix);
         } elseif ($element['raw'] ?? false) {
-            return !empty($element['route']) ? $element['route'] : '#';
+            return ! empty($element['route']) ? $element['route'] : '#';
         }
 
-        return !empty($element['route']) ? route($element['route'], $element['params'] ?? []) : '#';
+        return ! empty($element['route']) ? route($element['route'], $element['params'] ?? []) : '#';
     }
 }
 
-if (!function_exists('isActiveNavigation')) {
-    /**
-     * @param array $navigationElement
-     * @param string $navigationKey
-     * @param string $activeNavigationKey
-     * @return bool
-     */
-    function isActiveNavigation($navigationElement, $navigationKey, $activeNavigationKey)
+if (! function_exists('isActiveNavigation')) {
+    function isActiveNavigation(array $navigationElement, string $navigationKey, string $activeNavigationKey): bool
     {
-        $keysAreMatching = isset($activeNavigationKey) && $navigationKey === $activeNavigationKey;
+        $keysAreMatching = $navigationKey === $activeNavigationKey;
 
         if ($keysAreMatching) {
             return true;
