@@ -5,11 +5,15 @@ namespace A17\Twill\Http\Controllers\Admin;
 use A17\Twill\Services\Blocks\BlockMaker;
 use Illuminate\Filesystem\Filesystem;
 use Illuminate\Support\Collection;
-use Str;
+use Illuminate\Support\Str;
 use Symfony\Component\Finder\SplFileInfo;
 
 class IconsController extends Controller
 {
+    protected Filesystem $files;
+
+    protected BlockMaker $blockMaker;
+
     public function __construct(Filesystem $files, BlockMaker $blockMaker)
     {
         parent::__construct();
@@ -23,7 +27,7 @@ class IconsController extends Controller
         $icons = collect(
             config('twill.block_editor.directories.source.icons')
         )->reduce(function (Collection $keep, $path) {
-            if (!$this->files->exists($path)) {
+            if (! $this->files->exists($path)) {
                 return $keep;
             }
 
@@ -52,13 +56,12 @@ class IconsController extends Controller
     {
         $file = $this->blockMaker->getIconFile($file, false);
 
-        if (!$this->files->exists($file)) {
+        if (! $this->files->exists($file)) {
             abort(404);
         }
 
         return response()->stream(function () use ($file) {
             echo $this->files->get($file);
-        }, 200, ["Content-Type" => "image/svg+xml"]);
-
+        }, 200, ['Content-Type' => 'image/svg+xml']);
     }
 }
