@@ -13,22 +13,21 @@ use Illuminate\Http\Request;
 use Illuminate\Routing\Redirector;
 use Illuminate\Support\Facades\App;
 use Illuminate\View\Factory as ViewFactory;
+use Laravel\Socialite\Facades\Socialite;
 use PragmaRX\Google2FA\Google2FA;
-use Socialite;
 
+/**
+ |--------------------------------------------------------------------------
+ | Login Controller
+ |--------------------------------------------------------------------------
+ |
+ | This controller handles authenticating users for the application and
+ | redirecting them to your home screen. The controller uses a trait
+ | to conveniently provide its functionality to your applications.
+ |
+ */
 class LoginController extends Controller
 {
-    /*
-    |--------------------------------------------------------------------------
-    | Login Controller
-    |--------------------------------------------------------------------------
-    |
-    | This controller handles authenticating users for the application and
-    | redirecting them to your home screen. The controller uses a trait
-    | to conveniently provide its functionality to your applications.
-    |
-     */
-
     use AuthenticatesUsers;
 
     /**
@@ -157,7 +156,7 @@ class LoginController extends Controller
 
         $user = User::findOrFail($userId);
 
-        $valid = (new Google2FA)->verifyKey(
+        $valid = (new Google2FA())->verifyKey(
             $user->google_2fa_secret,
             $request->input('verify-code')
         );
@@ -193,19 +192,18 @@ class LoginController extends Controller
      */
     public function handleProviderCallback($provider, OauthRequest $request)
     {
-
         $oauthUser = Socialite::driver($provider)->user();
         $repository = App::make(UserRepository::class);
 
         // If the user with that email exists
         if ($user = $repository->oauthUser($oauthUser)) {
-
             // If that provider has been linked
             if ($repository->oauthIsUserLinked($oauthUser, $provider)) {
                 $user = $repository->oauthUpdateProvider($oauthUser, $provider);
 
                 // Login and redirect
                 $this->authManager->guard('twill_users')->login($user);
+
                 return $this->afterAuthentication($request, $user);
             } else {
                 if ($user->password) {
@@ -215,12 +213,14 @@ class LoginController extends Controller
                     $request->session()->put('oauth:user_id', $user->id);
                     $request->session()->put('oauth:user', $oauthUser);
                     $request->session()->put('oauth:provider', $provider);
+
                     return $this->redirector->to(route('admin.login.oauth.showPasswordForm'));
                 } else {
                     $user->linkProvider($oauthUser, $provider);
 
                     // Login and redirect
                     $this->authManager->guard('twill_users')->login($user);
+
                     return $this->afterAuthentication($request, $user);
                 }
             }
@@ -231,6 +231,7 @@ class LoginController extends Controller
 
             // Login and redirect
             $this->authManager->guard('twill_users')->login($user);
+
             return $this->redirector->intended($this->redirectTo);
         }
     }
