@@ -22,7 +22,12 @@ use PDO;
 
 abstract class ModuleRepository
 {
-    use HandleDates, HandleBrowsers, HandleRelatedBrowsers, HandleRepeaters, HandleFieldsGroups, HasCapsules;
+    use HandleDates;
+    use HandleBrowsers;
+    use HandleRelatedBrowsers;
+    use HandleRepeaters;
+    use HandleFieldsGroups;
+    use HasCapsules;
 
     /**
      * @var \A17\Twill\Models\Model
@@ -69,7 +74,7 @@ abstract class ModuleRepository
         $query = $this->filter($query, $scopes);
         $query = $this->order($query, $orders);
 
-        if (!$forcePagination && $this->model instanceof Sortable) {
+        if (! $forcePagination && $this->model instanceof Sortable) {
             return $query->ordered()->get();
         }
 
@@ -115,6 +120,7 @@ abstract class ModuleRepository
     public function getCountForAll()
     {
         $query = $this->model->newQuery();
+
         return $this->filter($query, $this->countScope)->count();
     }
 
@@ -124,6 +130,7 @@ abstract class ModuleRepository
     public function getCountForPublished()
     {
         $query = $this->model->newQuery();
+
         return $this->filter($query, $this->countScope)->published()->count();
     }
 
@@ -133,6 +140,7 @@ abstract class ModuleRepository
     public function getCountForDraft()
     {
         $query = $this->model->newQuery();
+
         return $this->filter($query, $this->countScope)->draft()->count();
     }
 
@@ -142,6 +150,7 @@ abstract class ModuleRepository
     public function getCountForTrash()
     {
         $query = $this->model->newQuery();
+
         return $this->filter($query, $this->countScope)->onlyTrashed()->count();
     }
 
@@ -173,7 +182,7 @@ abstract class ModuleRepository
 
         if ($this->model instanceof Sortable) {
             $query = $query->ordered();
-        } elseif (!empty($orders)) {
+        } elseif (! empty($orders)) {
             $query = $this->order($query, $orders);
         }
 
@@ -256,20 +265,17 @@ abstract class ModuleRepository
         return $this->hydrate($object, $fields);
     }
 
-    /**
-     * @param array $attributes
-     * @param array $fields
-     * @return \A17\Twill\Models\Model
-     */
-    public function updateOrCreate($attributes, $fields)
+    public function updateOrCreate(array $attributes, array $fields): Model
     {
         $object = $this->model->where($attributes)->first();
 
-        if (!$object) {
+        if (! $object) {
             return $this->create($fields);
         }
 
         $this->update($object->id, $fields);
+
+        return $object;
     }
 
     /**
@@ -335,6 +341,7 @@ abstract class ModuleRepository
             if (($object = $this->model->find($id)) != null) {
                 $object->update($values);
                 $this->afterUpdateBasic($object, $values);
+
                 return true;
             }
 
@@ -359,7 +366,6 @@ abstract class ModuleRepository
      */
     public function duplicate($id, $titleColumnKey = 'title')
     {
-
         if (($object = $this->model->find($id)) === null) {
             return false;
         }
@@ -393,11 +399,13 @@ abstract class ModuleRepository
                 return false;
             }
 
-            if (!method_exists($object, 'canDeleteSafely') || $object->canDeleteSafely()) {
+            if (! method_exists($object, 'canDeleteSafely') || $object->canDeleteSafely()) {
                 $object->delete();
                 $this->afterDelete($object);
+
                 return true;
             }
+
             return false;
         }, 3);
     }
@@ -418,6 +426,7 @@ abstract class ModuleRepository
                 if (config('app.debug')) {
                     throw $e;
                 }
+
                 return false;
             }
 
@@ -437,6 +446,7 @@ abstract class ModuleRepository
             } else {
                 $object->forceDelete();
                 $this->afterDelete($object);
+
                 return true;
             }
         }, 3);
@@ -460,6 +470,7 @@ abstract class ModuleRepository
                 });
             } catch (\Exception $e) {
                 Log::error($e);
+
                 return false;
             }
 
@@ -477,6 +488,7 @@ abstract class ModuleRepository
             if (($object = $this->model->withTrashed()->find($id)) != null) {
                 $object->restore();
                 $this->afterRestore($object);
+
                 return true;
             }
 
@@ -502,6 +514,7 @@ abstract class ModuleRepository
                 });
             } catch (\Exception $e) {
                 Log::error($e);
+
                 return false;
             }
 
@@ -518,11 +531,11 @@ abstract class ModuleRepository
     {
         if (property_exists($this->model, 'checkboxes')) {
             foreach ($this->model->checkboxes as $field) {
-                if (!$this->shouldIgnoreFieldBeforeSave($field)) {
-                    if (!isset($fields[$field])) {
+                if (! $this->shouldIgnoreFieldBeforeSave($field)) {
+                    if (! isset($fields[$field])) {
                         $fields[$field] = false;
                     } else {
-                        $fields[$field] = !empty($fields[$field]);
+                        $fields[$field] = ! empty($fields[$field]);
                     }
                 }
             }
@@ -530,14 +543,14 @@ abstract class ModuleRepository
 
         if (property_exists($this->model, 'nullable')) {
             foreach ($this->model->nullable as $field) {
-                if (!isset($fields[$field]) && !$this->shouldIgnoreFieldBeforeSave($field)) {
+                if (! isset($fields[$field]) && ! $this->shouldIgnoreFieldBeforeSave($field)) {
                     $fields[$field] = null;
                 }
             }
         }
 
         foreach ($fields as $key => $value) {
-            if (!$this->shouldIgnoreFieldBeforeSave($key)) {
+            if (! $this->shouldIgnoreFieldBeforeSave($key)) {
                 if (is_array($value) && empty($value)) {
                     $fields[$key] = null;
                 }
@@ -741,7 +754,7 @@ abstract class ModuleRepository
             }
 
             foreach ($object->$relationship as $relationshipObject) {
-                if (!in_array($relationshipObject->$attribute, $fields[$formField])) {
+                if (! in_array($relationshipObject->$attribute, $fields[$formField])) {
                     $relationshipObject->delete();
                 }
             }
@@ -801,7 +814,6 @@ abstract class ModuleRepository
      */
     public function searchIn($query, &$scopes, $scopeField, $orFields = [])
     {
-
         if (isset($scopes[$scopeField]) && is_string($scopes[$scopeField])) {
             $query->where(function ($query) use (&$scopes, $scopeField, $orFields) {
                 foreach ($orFields as $field) {
@@ -860,12 +872,12 @@ abstract class ModuleRepository
      */
     protected function getModelRepository($relation, $modelOrRepository = null)
     {
-        if (!$modelOrRepository) {
-            if (class_exists($relation) && (new $relation) instanceof Model) {
+        if (! $modelOrRepository) {
+            if (class_exists($relation) && (new $relation()) instanceof Model) {
                 $modelOrRepository = Str::afterLast($relation, '\\');
             } else {
                 $morphedModel = Relation::getMorphedModel($relation);
-                if (class_exists($morphedModel) && (new $morphedModel) instanceof Model) {
+                if (class_exists($morphedModel) && (new $morphedModel()) instanceof Model) {
                     $modelOrRepository = (new \ReflectionClass($morphedModel))->getShortName();
                 } else {
                     $modelOrRepository = ucfirst(Str::singular($relation));
@@ -880,7 +892,7 @@ abstract class ModuleRepository
         if ($repository instanceof ModuleRepository) {
             return $repository;
         } else {
-            $class = Config::get('twill.namespace') . "\\Repositories\\" . ucfirst($modelOrRepository) . "Repository";
+            $class = Config::get('twill.namespace') . '\\Repositories\\' . ucfirst($modelOrRepository) . 'Repository';
         }
 
         if (class_exists($class)) {
@@ -941,7 +953,7 @@ abstract class ModuleRepository
 
     /**
      * @param string $behavior
-     * @return boolean
+     * @return bool
      */
     public function hasBehavior($behavior)
     {
@@ -955,7 +967,7 @@ abstract class ModuleRepository
     }
 
     /**
-     * @return boolean
+     * @return bool
      */
     public function isTranslatable($column)
     {
