@@ -101,7 +101,7 @@ class BlockCollection extends Collection
      */
     public function readBlocks($directory, $source, $type = null)
     {
-        if (!$this->fileSystem->exists($directory)) {
+        if (! $this->fileSystem->exists($directory)) {
             $this->addMissingDirectory($directory);
 
             return collect();
@@ -159,10 +159,7 @@ class BlockCollection extends Collection
         return $block->source;
     }
 
-    /**
-     * @return $this
-     */
-    public function load()
+    public function load(): self
     {
         $this->generatePaths();
 
@@ -191,7 +188,7 @@ class BlockCollection extends Collection
         // remove duplicate Twill blocks
         $appBlocks = $this->collect()->whereIn('source', [Block::SOURCE_APP, Block::SOURCE_CUSTOM]);
         $this->items = $this->collect()->filter(function ($item) use ($appBlocks) {
-            return !$appBlocks->contains(function ($block) use ($item) {
+            return ! $appBlocks->contains(function ($block) use ($item) {
                 return $item->source === Block::SOURCE_TWILL && $item->name === $block->name;
             });
         })->values()->toArray();
@@ -204,15 +201,11 @@ class BlockCollection extends Collection
     }
 
     /**
-     * This function will add blocks and repeaters that are only defined in the config
+     * This function will add blocks and repeaters that are only defined in the config.
      *
      * For compatibility with 2.0.2 and lower
-     *
-     * @param Collection $items
-     * @param string $type
-     * @return void
      */
-    public function addBlocksFromConfig(Collection $items, $type)
+    public function addBlocksFromConfig(Collection $items, string $type): self
     {
         $items->reject(function ($value, $blockName) use ($type) {
             return $this->contains(function ($block) use ($blockName, $type) {
