@@ -29,7 +29,7 @@ class Imgix implements ImageServiceInterface
         $this->config = $config;
 
         $urlBuilder = new UrlBuilder(
-            $this->config->get('twill.imgix.source_host'),
+            $this->config->get('twill.imgix.source_host') ?? 'example.imgix.net',
             $this->config->get('twill.imgix.use_https'),
             '',
             false
@@ -52,7 +52,7 @@ class Imgix implements ImageServiceInterface
         $defaultParams = $this->config->get('twill.imgix.default_params');
         $addParamsToSvgs = config('twill.imgix.add_params_to_svgs', false);
 
-        if (!$addParamsToSvgs && Str::endsWith($id, '.svg')) {
+        if (! $addParamsToSvgs && Str::endsWith($id, '.svg')) {
             return $this->urlBuilder->createURL($id);
         }
 
@@ -158,6 +158,7 @@ class Imgix implements ImageServiceInterface
         } catch (\Exception $e) {
             try {
                 list($width, $height) = getimagesize($url);
+
                 return [
                     'width' => $width,
                     'height' => $height,
@@ -177,9 +178,8 @@ class Imgix implements ImageServiceInterface
      */
     protected function getCrop($crop_params)
     {
-        if (!empty($crop_params)) {
-            return ['rect' =>
-                $crop_params['crop_x'] . ',' .
+        if (! empty($crop_params)) {
+            return ['rect' => $crop_params['crop_x'] . ',' .
                 $crop_params['crop_y'] . ',' .
                 $crop_params['crop_w'] . ',' .
                 $crop_params['crop_h'],
@@ -197,7 +197,7 @@ class Imgix implements ImageServiceInterface
      */
     protected function getFocalPointCrop($crop_params, $width, $height)
     {
-        if (!empty($crop_params)) {
+        if (! empty($crop_params)) {
             // determine center coordinates of user crop and express it in term of original image width and height percentage
             $fpX = ($crop_params['crop_w'] / 2 + $crop_params['crop_x']) / $width;
             $fpY = ($crop_params['crop_h'] / 2 + $crop_params['crop_y']) / $height;
@@ -212,7 +212,7 @@ class Imgix implements ImageServiceInterface
             $params = ['fp-x' => $fpX, 'fp-y' => $fpY, 'fp-z' => $fpZ];
 
             return array_map(function ($param) {
-                return number_format($param, 4, ".", "");
+                return number_format($param, 4, '.', '');
             }, $params) + ['crop' => 'focalpoint', 'fit' => 'crop'];
         }
 
